requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
google-generativeai==0.8.5
httpx==0.28.1
python-dotenv==1.1.0
